requirements.txt
-f /usr/share/pip-wheels
backcall==0.2.0
boto3==1.16.30
botocore==1.19.30
click==7.1.2
decorator==4.4.2
Flask==1.1.2
Flask-SQLAlchemy==2.4.4
ipykernel==5.1.3
ipython==7.19.0
ipython-genutils==0.2.0
itsdangerous==1.1.0
jedi==0.17.2
Jinja2==2.11.2
jmespath==0.10.0
jupyter-client==6.1.7
jupyter-core==4.7.0
MarkupSafe==1.1.1
numpy==1.19.2
opencv-python==4.4.0.46
pandas==1.1.3
parso==0.7.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.0.1
prompt-toolkit==2.0.10
ptyprocess==0.6.0
Pygments==2.7.3
python-dateutil==2.8.1
python-dotenv==0.15.0
pytz==2020.4
pyzmq==20.0.0
s3transfer==0.3.3
six==1.15.0
SQLAlchemy==1.3.20
torch==1.4.0+cpu
torchvision==0.5.0+cpu
tornado==6.0.3
traitlets==5.0.5
urllib3==1.26.2
wcwidth==0.2.5
Werkzeug==1.0.1
python-dotenv
